+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 
+## 0.0.2 (Jun 22, 2015)
+
+- Support a shortcut for showing message content.
+- Improve the `options` argument of "show" method.
+
+
 ## 0.0.1 (Jun 10, 2015)
 
 - Supports eleven options: "autohide", "autoshow", "content", "countdown", "dismissible", "duration", "fixed", "offset", "position", "style", "zIndex"